Enable image lazy loading in file manager (see #3815)
Description
-----------

Opening the file manager while a folder is open that contains a large amount of images can cause a lot of load on the web server since the browser will try to load all images immediately and the server will then either serve the images directly - or still has to generate the images, if not already done (if the image cache was cleared previously for instance).

Contao 4.9 supports deferred resizing and thus we can take full advantage of `loading="lazy"` here. I consider it a bug that we did not add this alongside https://github.com/contao/contao/pull/703 back then.

Commits
-------

7c0c4ed9 enable image lazy loading in file manager