Rework the @throws annotations (see #3816)
Description
-----------

With regard to #3791, this is a new attempt to create rules for `@throws` annotations.

As already discussed, we do not want to add `@throws` exceptions to every method that calls another method that calls another method that might throw an exception. Instead, we want to add `@throws` to indicate that an exception should be handled in a try/catch block in contrast to letting it bubble up to the general exception handler.

### Example 1

```php
/**
 * @throws \InvalidArgumentException If the file does not exist
 */
public function loadFile(string $file): SplFileInfo
{
    // ...
}
```

We add the `@throws` annotation here to indicate that the method should not be used without a try/catch block.

### Example 2

```php
public function __construct(string $func, string $action = null, string $username = null)
{
    if ('' === $func) {
        throw new \InvalidArgumentException('The function name in the Contao context must not be empty');
    }

    // ...
}
```

We do not add a `@throws` annotation here because we want the exception to bubble up to the general exception listener. The developer is not supposed to add a try/catch block here but to provide a non empty function name.

### Conclusions

Following these rules, `@throws` annotations never make sense for

- private methods (because only we use them);
- internal classes (e.g. event listeners);
- classes that nobody is supposed to work with (e.g. controllers);
- any exception that is supposed to bubble up to the general exception handler.

They primarily make sense for

- interfaces and abstract classes;
- methods that should only be used inside a try/catch block.

### Ambiguous cases

How about invalid argument exceptions? Are these meant to bubble up or is a developer supposed to wrap them in a try/catch block?

Commits
-------

f4e69076 Remove the annotation from private methods
538df1e1 Remove the annotation from the listeners
58accec3 Remove the annotation from the controllers
b6715c20 Remove exceptions that are supposed to bubble up
3531723c CS
4b170582 Incorporate the changes discussed in the PR