Add front end preview links (see #3872)
Description
-----------

This PR replaces the maintenance login feature _not_ introduced in #3729. As mentioned by @fritzmg in https://github.com/contao/contao/pull/3729#issuecomment-999468916, a regular use-case is not only to bypass the maintenance mode, but also to show unpublished pages to users that do not have a back end account (e.g. management to approve a new website or page).

This PR introduces _frontend preview links_. It work using a signed (= secure) link that logs in the front end user with our existing `FrontendPreviewToken`, which then allows to see pages in maintenance and optionally unpublished elements, even if not logged in to the back end.

I have discussed the feature with @Toflar and @netzarbeiter and this is the best approach we could think of. For a user, you simply navigate to the desired page in the front end, and click the _Share URL_ button in the preview bar. This opens the back end to share that page. 

We initially discussed a checkbox to show or hide the preview bar in the front end as well. I however noticed that the preview bar only ever works if a backend user is logged in. Therefore this additional check/option is unnecessary, the preview bar is simply not available for regular visitors through preview links. If they are for any reason logged in to the back end, the preview bar will appear as usual.

Here's a quick demo:

https://user-images.githubusercontent.com/1073273/147571764-87962ace-f0f6-4b08-b928-17268359a593.mov


### Security implications:
As you might notice in the PR, there are a few places where we currently check for `isGranted('ROLE_USER')` **in addition to** checking for `FrontendPreviewToken`. This is (a) inconsistent because we don't always to that and (b) obviously wouldn't work with this PR.
 - the `FrontendPreviewToken` is currently only ever set in the preview bar controller, and that controller double-checks on `ROLE_USER` permissions.
 - theoretically, a user could currently "loose" its backend firewall token and would then not have frontend preview either. I don't really see this ever happening though, even a logout of the back end invalidates the session and logs out any front end users.

### TODO:
- [x] Move DCA callbacks to event listener
- [x] Implement sharing link (copy to clipboard) in the DCA list view
- [x] Update the `share.svg` icon to reflect adding to clipboard
- [x] Unit tests

Commits
-------

4c342260 Introducing frontend preview links
a191962a Adjust legacy classes to not check backend user with frontend preview
91d4c5a7 Correctly enable front end preview if unpublished are not shown
00d8b313 Merge remote-tracking branch 'upstream/4.x' into feature/preview-links
cbebd8e1 Renamed table and added listener for callbacks
b8c8f95b Fixed and implemented toggle button
3bd45275 Updated test for preview toolbar
aa46adee Fixed tests & CS
03a00412 Updated tests
22e3b056 CS
7ae5532f CS
fb3d4ea5 More tests
93ccc820 CS
f2b2b011 CS
3378c985 Fixed field name in toggle operation
31a648eb Added tests for PreviewLinkListener
cf5a2dba CS
e7413654 CS
4e537073 Use toggle operation and implement permission checks
cd9bcdbb Fixed tests
83a2514e CS
7de1cc1c More tests for PreviewLinkListener
92594377 Updated the share icon
c0bd42e7 Encode the ampersand character
f799c788 CS
f0ce4e1a Adjust the labels
0f47c379 Adjust the palette and the CSS for readonly fields
1484e3bb Remove the JS clipboard confirmation

Co-authored-by: Leo Feyer <github@contao.org>