Add options to customize the layout inheritance for pages (see #3419)
Description
-----------

Add an option to configure the template inheritance from parent pages to child pages.

Possible options are:
- use layout on subpages (current behaviour, default & fallback behaviour)
- use layout on this page only
- use a custom layout for subpages

> <img width="771" alt="contao-layout-inheritance" src="https://user-images.githubusercontent.com/1193129/132098890-71aefb94-b801-44cf-a846-5edff2c0c95a.png">

It is much easier to implement special overview pages or more complex layout structures via the page structure in this way:
- If a single overview page requires a special layout, define `use layout on this page only` – subpages then use the layout of the overview page’s parents (the current page is skipped when walking through the parents looking for a layout).
- If all subpages of a page should have the same layout which is not the layout of the parent page, set `use a custom layout for subpages` and select the layout for the subpages.

Overall, this decouples layout inheritance somewhat from the page tree (which is primarily used for menu structure).
Another advantage is that the default inheritance of the layout to subpages is displayed directly to editors by the default  option `use layout on subpages`.


**Handling missing configuration:**
This new implementation still does not resolve a bad/missing configuration:
- If a parent page has `includeLayout` set but has no `layout`, the empty `layout` is inherited by the current page.
- If a parent page has `layoutInheritance` set to `use a custom layout for subpages` but has no `subpagesLayout`, the empty `subpagesLayout` is inherited by the current page.

Since the layout select widgets do not have a blank option, both cases should not happen when editing a page mindfully, but could happen after a `submitOnChange` without a manual save after.

One could argue that in case of a missing layout value, the layout should not be inherited from this parent.


**Target Contao version:**
As a feature, this should probably be added to the latest Contao version. But since it is a progressive enhancement which defaults to the current behavious, it could be added to Contao 4.9 rather easily. The current pull request is based on Contao 4.9.

Issues on this topic:
- #1797
- https://github.com/contao/core/issues/1270 (very old)
- https://github.com/contao/core/issues/4868 (very old)

Commits
-------

605b3eb9 add options to customize layout inheritance
938cfc02 remove tl_page.layoutInheritance
7b6b5691 add PageSubpagesLayoutMigration
bd4ccf6a fix note in UPGRADE.md
2e810b03 add tl_page.subpagesLayout to functional fixtures
19097318 Harden migration
d58c5f55 Remove side effect from PageModelTest
d333de80 add foreignOptions configuration for DCA widgets
0195df23 fix yaml in functional test fixtures
0f10814a cs fixes
d9756145 add missing type-hint
e09a94a9 remove tl_page.layoutPropagation & foreignOptions
823539dd Remove left over subpalette
080503ce add LayoutOptionsListener
ad9dbb73 implement ResetInterface for LayoutOptionsListener
85a01953 Rename "subpages layout" to "subpage layout"
10744a5b Rename "subpages layout" to "subpage layout"
15859d70 fix code style

Co-authored-by: Leo Feyer <github@contao.org>