"Warning: Attempt to read property "language" on null" with PHP8
I am getting a warning in backend with debug mode enabled when I open an article containing the downloads element.

![image](https://user-images.githubusercontent.com/87128053/150656068-0c2c4eb2-8826-49e2-9825-e0f4471b3d7e.png)

The problem is, that $objPage is NULL in backend. When I tried to access a property from NULL PHP7 generated just a notice that was ignored in the debug mode. PHP8 generates a warning, so I have to disable the debug mode when I try to open a article containing a downloads element in backend.

I changed the code to check, if $objPage exists.