Fix long content being cut off in the back end popup (see #4009)
Description
-----------

When viewing the record information in a popup, it may happen that the data without whitespaces (such as JSON) is cut off at the right. This pull request fixes this problem and makes the popup content container scrollable horizontally.

![2022-01-26 at 15](https://user-images.githubusercontent.com/193483/151177730-54c4c41a-8c0c-44b8-be13-b782bff54ff8.png)

/cc @Toflar

Commits
-------

c4c9e64e Fix the long content being cut off in the backend popup when viewing the record information