Fix the missing request token in ModulePassword.php (see #4060)
Description
-----------

There is a bug in the lostPassword module since Conato 4.13.

The requestToken is missing, when I open the link in the mail to change the password.

This is because when I open the url to change the password with the get parameter "token", the script doesnt reach the token definition on line 177.

https://github.com/contao/contao/blob/3691e1d7f9ee4eb35a9f5e05a3537dc633f94ffc/core-bundle/src/Resources/contao/modules/ModulePassword.php#L77-L83

https://github.com/contao/contao/blob/3691e1d7f9ee4eb35a9f5e05a3537dc633f94ffc/core-bundle/src/Resources/contao/modules/ModulePassword.php#L177

I changed the file ModulePassword.php to set the requestToken above the if statement.

Commits
-------

2eb56d0e Fixed missing requestToken in ModulePassword.php