Correctly handle null values in DCs (see #4067)
Description
-----------

With #3876 we introduced a bug with `null` values because `is_scalar(null) === false` in PHP. 🙈

This means that `eval.doNotSaveEmpty` does not work anymore for `null` values and now tries to save them to the database.

This breaks our custom elements extension because it uses `doNotSaveEmpty` for its “virtual” fields. 🥺

Commits
-------

9059c9b6 Correctly handle null values in DCs