Support virtual filesystem in CLI backup management  (see #4004)
Description
-----------

This ~outlines~/implements the changes required to support the new VFS in our (also new) backup management.

It's a demonstration as to how easy it is to work with the new VFS, thank you @m-vo <3 

Implementing it also showed a few minor flaws in the current backup design. Like `Backup` knowing about it's physical location as well as size which should be the `BackupManager`'s responsibility.

Certainly, changing to the VFS is something for Contao 5 or later but we might think about fixing the responsiblity stuff and removing the configuration parameter for `var/backups` in Contao 4.13 before the release. I think it would be better to introduce VFS support where you can configure the mount to an S3 running on cloud computers on the moon if you like rather than having a configurable local path right now.

On the other hand, we can also leave 4.13 as 5 will allow us to break also configuration.
Also depends a bit on @aschempp if he wants to introduce backup support in the Contao Manager. In the current implementation the full local path to the backup is returned in `contao:backup:list --format=json` so he could offer the backup for download. Reading from a local path obviously won't be possible anymore if the backup directory is mounted to S3 for example. That's why I included a `contao:backup:stream-content <backup-name>` command in this PR. It allows to stream and thus offer the backup contents via CLI and then the Manager no matter where they reside.

Commits
-------

472657a2 Reworked CLI backup management to support the virtual file system
98fe7572 Update core-bundle/src/Doctrine/Backup/BackupManager.php
f1da7d70 Hide contao:backup:stream-content from the CLI
e1c1f1c5 Rename variable
d48e7f25 Adjust to changes in the VFS
e8095afa Make contao:backup:list command aware of server time
149f53a9 Make unit tests run in non-UTC timezones
575dc2f3 Fixed localizing backup datetime