Do not configure the default click jacking paths in the skeleton (see #4135)
Description
-----------

Suppose you want to allow a certain URL (or URLs) to be used as an `<iframe>` on another domain. What you would normally do is the following in your `config/config.yaml`:

```yaml
nelmio_security:
    clickjacking:
        paths:
            '^/foobar/.+': ALLOW
            '^/.*': SAMEORIGIN
```

However, this does not actually work. The reason is the loading of the `skeleton/config.yml` of the Manager Bundle. It messes up the order of paths, causing `'^/.*': SAMEORIGIN` to always be the first element in the resulting config array after the configurations were processed by Symfony. Therefore the `X-Frame-Options` response header will always contain `SAMEORIGIN` and changing that header for a specific URL will have no effect - you can only set it to `ALLOW` globally.

To fix this, the default `'^/.*': SAMEORIGIN` value (needed for the Contao back end) should only be set via `ExtensionPluginInterface`, if the rule for that path was not defined yet.

Commits
-------

2fb6686b do not override clickjacking paths in skeleton
d839298b code style
462fbf77 do not pass container
8b158837 Adjust the phpDoc comment

Co-authored-by: Leo Feyer <github@contao.org>