Unset key in URL when switching edit mode (see #4212)
Description
-----------

I'm using a custom global operation and backend module (`$GLOBALS['BE_MOD'][...]['foo'] = ['controller', 'method')` to create a database record with specific data. I can leverage the `DataContainer::create` method for that, which explicitly allows to pass additional data to the new records.

When creating the data, `DataContainer::create` redirects to the new record and in the process removes known "invalid" parameters from the URL. But since it does not remove `key`, it results in an endless redirect (because my module is called again). I think we can safely remove `key` as well since thats a well supported pattern imho.

Commits
-------

1fd9b982 Unset key in URL when switching edit mode