Drop the pageSelector and fileSelector widgets (see #3993)
Description
-----------

These have not been in use for ages. Be aware that the `fileTree` and `pageTree` are different! Even though some of their methods say they are deprecated, that's actually what we still use everywhere, and changing that would be a different PR.

PS: this is a (small) first step in preparing to rewrite our backend JS. We need to get rid of unnecessary stuff before a reasonable attempt can be made. So it would be good to get this merged sooner than later 😊

Commits
-------

7df455fc Drop the pageSelector and fileSelector widgets
d363f166 Fix backend path rewrite