Add a warning for too large database row sizes (see #4179)
Description
-----------

To improve the situation for #4159 I created a small tool that is able to measure and calculate row sizes.

Not sure if this should ever get merged, but I needed a way to write the (somewhat confusing) rules for row sizes down and be able to validate them.

`getMysqlColumnSizeBits()` and `getInnodbColumnSizeBits()` calculates the amout of bits the column contributes to the row size.

`measureMysqlColumnSizeBits()` and `measureInnodbColumnSizeBits()` return the exact same information, but evaluate the amout of bits needed by trail and error against the database. (slow, only used for testing)

![Bildschirmfoto 2022-02-19 um 23 46 42](https://user-images.githubusercontent.com/367169/154822307-81215715-14db-4156-a683-367eb2ef0ceb.png)

Commits
-------

f672abd3 Add a warning for too large database row sizes
92cba100 Fix column charset detection
d4d4858c Reduce row size of tl_content
7f0d8ce4 Use customSchemaOptions
8db72fa4 Reduce table row size for tl_content and tl_module
6b2d8f47 Merge remote-tracking branch 'origin/4.13' into fix/row-size-too-large
16b60a79 Increase image size length
467c8d07 Increase image size length
43dfbb33 Support custom collations in the DcaSchemaProvider

Co-authored-by: Leo Feyer <github@contao.org>