Ignore empy layout selection (see #5730)
Description
-----------

Fixes #5729

Currently, as soon as you click on **Assign a layout** in the settings of the page, that page will not work anymore in the front end, with the error "No layout specified". This is because the layout assignment is activated but no layout has been set yet at that point (assuming no layout was set previously in the database for that page).

This however can easily be fixed by letting `PageModel::loadDetails` ignore the `includeLayout` setting, if the `layout` setting is still falsey (i.e. `0`).

Commits
-------

cdc76011 ignore empty layout selection
c7491b6e Merge branch '4.9' into fix-no-layout-specified