Always set `currentRecord` when initializing widgets (see #5753)
Description
-----------

First of all, this is **not to be mistaken** with `$objCurrentRecord` in the DCA, which we have recently reworked for Contao 5.

Our widgets have a `currentRecord` property, which is set when they are initialized:

https://github.com/contao/contao/blob/e581d92fc8d7d2447369acf036265b41ad567852/core-bundle/src/Resources/contao/classes/DataContainer.php#L433-L435

However, we are not doing this everywhere, so sometimes the value is unexpectedly `null`. This PR fixes that.

Commits
-------

04b8d3ad Always set `currentRecord` when initializing widgets