Improve handling of pages with `tl_page.requireItem` (see #3540)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #3450 , #3525 

With the new routing, trying to generate a URL for a page with `tl_page.requireItem=1` throws a `RouteParametersException`. This pull request prevents or improves the handling of this exception in breadcrumbs and navigation modules as well as in the frontend preview.

Improved cases:

**1. Breadcrumbs, current item has `requireItem`**
Generate the URL from the current request instead of re-generating it from the taget page.

Note: The URL is generated via `strtok(Environment::get('request'), '?')` so any query string is not added to the breadcrumbs URL. 

**2. Breadcrumbs, item in parent trail has `requireItem`**
Skip pages with `tl_page.requireItem=1` in the parent trail of the breadcrumbs, even if `showHidden` is enabled.

Note: These pages are skipped explicitly in the breadcrumbs (and not handled by the future proof fix below) to prevent spamming the log with error messages for failed URL generations.

**3. Future proof breadcrumbs**
Only add items with a non-empty URL (`$href`) to the breadcrumbs (except root page).

This way, if some page types will not generate a proper URL (after an exception or for pages with no public route in the future), they will neither get added to the breadcrumbs nor to the JSON-LD data.

**4. Frontend preview**
If a `RouteParametersException` is thrown for a page when converting the frontend preview URL and the reason for the exception is an `InvalidParameterException` on a page with `tl_page.requireItem`, catch the exception and do not set a URL for the preview, thus starting the preview on the root page.

Commits
-------

a004075a improve handling of pages with requireItem
2f01408d Do not create preview link for pages with requireItem
a5c2822b Merge branch '4.13' into fix/pages-with-require-item
b400de71 CS
f77c09fa Fix tests
8257c94a CS
24e72164 Merge branch '4.13' into fix/pages-with-require-item
4cd350f3 CS

Co-authored-by: Andreas Schempp <andreas.schempp@terminal42.ch>
Co-authored-by: Leo Feyer <1192057+leofeyer@users.noreply.github.com>