Remove the `q` shortcut from `contao:crawl` (see #5879)
Description
-----------

In #5795 we added the `--queue` option to the `contao:crawl` command. However, we also defined the shortcut `q` for this option - which unfortunately is already used for the global `--quiet` option, thus resulting in the following error:

```
$ vendor/bin/contao-console contao:crawl --help
18:32:59 CRITICAL  [console] Error thrown while running command ""contao:crawl" --help". Message: "An option with shortcut "q" already exists." ["exception" => Symfony\Component\Console\Exception\LogicException^ { …},"command" => ""contao:crawl" --help","message" => "An option with shortcut "q" already exists."]  

  An option with shortcut "q" already exists.
```

This also breaks the Contao Manager since this error also occurs when running `list --format=json`:

```
$ vendor/bin/contao-console list --format=json
18:34:55 CRITICAL  [console] Error thrown while running command "list --format=json". Message: "An option with shortcut "q" already exists." ["exception" => Symfony\Component\Console\Exception\LogicException^ { …},"command" => "list --format=json","message" => "An option with shortcut "q" already exists."]        

  An option with shortcut "q" already exists.
```

This PR removes the shortcut from the option.

This will likely fix https://github.com/contao/contao-manager/issues/767 (and related issues with the Contao Manager).

Commits
-------

975aa015 remove the q shortcut from contao:crawl
a6032518 fix tests