Always allow toggling a field that is not excluded (see #5876)
Description
-----------

I have a table where none of the fields are excluded since there's no reason to set user permissions on individual fields. Regular users do not see the toggle icon now, and they can't call the toggle operation.

Be aware that this behaviour has changed in Contao 5, so this needs adjustment to merge upstream. I'm happy to provide a Contao 5 PR once this is agreed/merged 😊

Commits
-------

eebd6808 Always allow toggling a field that is not excluded
63788d52 Fixed variable for table