Consider foreign keys when sorting the list view (see #5917)
Description
-----------

Currently, the foreignKey is not taken into account when sorting the listView. With this adjustment, if a field uses a foreignKey, the values of the target table are automatically used for sorting.

Regardless of the foreignKey, usually only the name of the column is used for sorting. The adjustment inserts a SELECT instead of the column name.

```php
if (isset($GLOBALS['TL_DCA'][$this->strTable]['fields'][$key]['foreignKey']))
{
	$chunks = explode('.', $GLOBALS['TL_DCA'][$this->strTable]['fields'][$key]['foreignKey'], 2);
	$orderBy[$k] = "(SELECT " . Database::quoteIdentifier($chunks[1]) . " FROM " . $chunks[0] . " WHERE " . $chunks[0] . ".id=" . $this->strTable . "." . $key . ")";
}
```

The code was also minimally adapted to avoid redundancies.

Commits
-------

0a5dc232 Consider foreignKey when sorting listView
b5a035f9 Update DC_Table.php
5501de10 Update DC_Table.php
95652fd0 Update DC_Table.php
17c367a3 Update DC_Table.php