Use the title instead of the filename in the download component (see #5981)
Description
-----------

Fixes #5975

Although this can be easily adjusted in the template, the change wasn't intended. Now, if a text wasn't explicitly specified for an element, we default to the title. 

So a with a derived title `Lunch Menu`, you'll get something like `<a href="…" title="Download Lunch Menu">Lunch Menu</a>` instead of `<a href="…" title="Download Lunch Menu">lunch_menu_final_version1(1).pdf</a>`.

Note, that the file name will still be visible when downloading and should therefore be part of the editorial work.

Commits
-------

3178c4df fix download text defaulting to the file name