Ensure that root pages are always shown in the correct order (see #5966)
Description
-----------

I'm having an issue with the page tree displaying the root pages in a different order for a restricted user group rather than the admins. This is because we call `updateRoot()` at all sorts of different places where the array is not always sorted by `sorting`. I've fixed this now by always ordering within `updateRoot()` in case there is a `sorting` field and we do not have any `visibleRootTrails`. Because if we do, it means `$this->root` contains an ID which is not on top-level in which case we cannot order by `sorting` anymore - this would result in a wrong order as well.

So if your user group has page mounts that only have root pages selected, you now always have the same order as the admins do - guaranteed.

Commits
-------

4d6d25f9 Try to improve root order
b394b9a6 Update core-bundle/src/Resources/contao/drivers/DC_Table.php