Load the page details before manipulating root page data (see #6038)
Description
-----------

While updating a website from Contao 4.9 to 4.13 with multiple root pages, I had issues changing the route prefix once `legacy_routing` was disabled. It always told me a change in prefix or suffix would result in a conflict with an existing page, whos URL was not reflecting the just-changed prefix.

If page details are not loaded when changing the settings inherited from the root page, the changes will be overridden when page details are automatically loaded by the routing system.

Commits
-------

733be537 Load the page details before manipulating root page data