Always use `System::importStatic()` for callbacks and hooks (see #6163)
Description
-----------

This is a first step to get rid of the `System::import()` method, which I hope everyone agrees that we should. 😉 

Of course we have to keep known class properties such as `$this->Database` or `$this->User` to remain backwards compatible, so this pull request only handles dynamic class properties. The next step would be to replace all the remaining usages and to deprecate the method. I will create a follow-up pull request for that.

Commits
-------

b6122b26 Always use `System::importStatic()` for callbacks and hooks
9867e76f Use `System::importStatic()` in the maintenance module
3a695f06 Use `System::importStatic()` for button callbacks