Add end-to-end tests with Cypress (see #5465)
Description
-----------

Fixes #5273

The test requires an empty Contao installation as described in the [development section](https://github.com/contao/contao/blob/5.x/README.md#development).

For the backend login we run contao:migrate and create a user. Then the tests are executed. After the tests, the database is reset and ready for new tests.

Commits
-------

75652501 add backend login test

Co-authored-by:  <arzberger@pdir.de>