Skip type error when trying to assign DCA label to string (see #6225)
Description
-----------

I got the following error in my unit test setup, because `$v` was a string and not an array.

```
TypeError: Cannot access offset of type string on string
```

This most likely happened because the `DefaultOperationsListener` did not convert a DCA operation into an array (likely `tl_page.list.operations.show`). However, I still think this error should be skipped if it happens for any reason.

Commits
-------

00bca46c Skip type error when trying to assign DCA label to string