Automatically translate the default maintenance template (see #6494)
Description
-----------

Here's a solution that covers the 99% case. Most people won't adjust the maintenance template and so we can use JS to translate it according to the user's browser language if we have translations available.

So for German, this now looks like this automatically:

<img width="850" alt="Bildschirmfoto 2023-11-07 um 17 24 20" src="https://github.com/contao/contao/assets/481937/f586f6f4-9158-4b6f-a0c4-d69e1831964e">

We might improve this even more in the future by using the upcoming Symfony translation support for JS but this is still experimental and I don't know if this is really a good fit anyway. The reason for that is that it would tie the maintenance mode to require a running system again.
I know that we are already sort of relying on the fact that the container still works as we're using Twig when using `contao:maintenance enable`. So I guess this is kind of meant to work like "in case the System does work, we can use your template, if not, then you have to copy something to `var/maintenance.html` yourself".

So Twig should be fine. But using dynamic translations would probably be a bit too much.

PS: Translating the thing in the footer doesn't work because of placeholders. So this here is really KISS.

Commits
-------

0d205eac Automatically translate the default maintenance template
a56d8e74 Reuse translation method
674ee438 Update core-bundle/templates/Error/service_unavailable.html.twig
734ac9d0 Update manager-bundle/src/Command/MaintenanceModeCommand.php
a12bca51 Use const instead of let

Co-authored-by: leofeyer <github@contao.org>